修复供应商配置解析校验
